requirements.txt
fastapi
uvicorn
passlib[bcrypt]
requests
pytest
python-jose
csv
